remove user settings "Quick actions"
There were 3 "Quick actions": adding a quick link (which is a deprecated feature), adding a "search snippet" (a 3-year-old name for a rarely used feature), and changing the number of search context lines. Only the latter is arguably useful, but even that is rarely used. The other 2 are not useful. I don't think that the UI complexity is worth it for just the single quick action that is not old/deprecated.