remove FeedbackPrompt
The feedback prompt was shown in the top right in the global navbar.

While I like the idea of an easy feedback widget, this one is currently worse than nothing because we don't monitor it or have the ability to effectively respond:

- Submissions are anonymous and do not include the username or email address of the user. We can't easily get in touch with them in the current form.
- Submissions are not being responded to. Submissions go to productboard, but it gets lost there, and I don't see any recent responses to these.

It is also somewhat duplicative of the periodic survey that users see every 30 days unless they dismiss it.

History: This was added 2 years ago. It has only received general/maintenance fixes since then (such as accessibility and global refactors).

(If we want to revive it later when we can give it some attention, we can always revert this commit easily.)