remove code for self-service payments and billing
This has been deprecated for more than a year. We do not support self-service payment. See https://about.sourcegraph/pricing if you are a customer or want to be a customer.

The functionality remains for users to view license keys on Sourcegraph.com, and for us to generate license keys. There is no longer any way to see the billing history of subscriptions, but we have not supported the creation of such subscriptions for 2+ years. Any customers needing historical invoices may contact us.