remove code intel hover alerts
The HoverAlerts for the code navigation precision level are large and distracting. I would guess most users tune them out, and most users can’t or don’t take action based on them. Now that we are moving toward auto-indexing and accept that the burden is on us to make it easy for precise code navigation to be rolled out and on admins to actually set that up, they are even less valuable to show to each user.

This does not affect the hover alerts for the browser extension that inform the user that (eg) GitHub's native tooltips are being suppressed.