allow `git shortlog --after` to fix contributors time ranges
The repository contributors page relies on `git shortlog --after=...`, but the `--after` command-line option was not in the allowlist. This commit adds it there. This is safe because the `--after` (and `--before`) string is just interpreted as a date and can't have bad side effects; it's also allowed for `git log`, where it has the same meaning.

Fixes https://github.com/sourcegraph/sourcegraph/issues/41825.