remove needless ?toast=integrations from post-sign-in redirect URL (#42044)
We used to use this to trigger display of a toast to recommend users to install our browser extension. We no longer show that because it was annoying and users didn't like it. For some reason, the URL after sign-in or signup had ?toast=integrations manually added to it. This didn't do anything, and there is no reason to keep it.