prevent extraneous props from being applied to <svg>
Fixes devtools console errors on `/insights/about`:

```
react-dom.development.js:86 Warning: React does not recognize the `groupSearch` prop on a DOM element. If you intentionally want it to appear in the DOM as a custom attribute, spell it as lowercase `groupsearch` instead. If you accidentally passed it from a parent component, remove it from the DOM element.
```