worker: Move commandLogger out of executor
This might be useful elsewhere, to demo it I added a bunch of logs to the repo syncer. It's relatively easy to expose these in the UI, there's a unmarshal helper for the DB and a helper to create a GQL resolver for the ExecutionLogEntry type. To do a quick PoC, I skipped this part.