Make fuzzy finder modal wide again (#41736)
* Make fuzzy finder modal wide again

The width of the fuzzy finder modal recently became very narrow. I
wasn't able to track down the exact PR that impacted the width, but at
some point the priority of the Wildcard `Modal` component overwrote
the custom `width: 80vw` style we defined specifically for the fuzzy
finder. This commit fixes the priority by use an `id` attribute instead
of `class`.

* web: make fuzzy finder modal wide again

Co-authored-by: Valery Bugakov <skymk1@gmail.com>