codeintel: reduce search based definition query count to 50 (#42224)
When we disabled legacy extensions on Sourcegraph.com (around 13:40 20th
Sep UTC), Zoekt's CPU usage and contention went up significantly. In
code-intel-extensions repository, the definition query uses count:50.
However, in the Sourcegraph repo we use count:500. We believe this is
the root cause for the regression.

Using the bubble up feature on honeycomb, it shows that the difference
between our bad queries and good is a max shard match count of 500.

Test Plan: CI and then monitoring CPU usage in production.

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>