database: Don't allow empty code host config (#41985)
We already have a db constraint to stop this in most cases, but when we
have encryption enabled then even an empty config is converted into an
encrypted value so we need to do the check in our application layer too.

While investigating this we already had code to validate our config but
it had not been included in the Upsert path. That is now fixed and tests
were updated where necessary.