References Panel: promote on second click/enter
Previously, it was necessary to move the mouse to the preview pane to
promote a line. This PR adds the ability to promote the "active" item by
clicking on it a second time (or pressing enter).

Demo https://www.loom.com/share/1d7aa783fcc84387b82a5c6443c0bf11