CodeMirror blob: Override browser search to trigger file search (#42255)
File pages now override the browser's search trigger. CTRL/CMD + f will
trigger CodeMirror's document search instead of the browser's built-in
search. Because CodeMirror doesn't necessarily render all lines of a
document it's not possible to user the browser's built-int search to
search inside the file.

Additional changes in this PR (which I can move in a separate PR if
desired):

- Introduced a special `'Mod'` modifier that represents Command/Meta on
  macOS and `Ctrl` elsewhere.
- Added a utililty function to set the platform value in unit tests.