CodeMirror blob: Show pointer cursor when "click to go to def/ref" is available
The logic for showing a pointer cursor in the old blob view resides in
the "click to go to definition" logic inside `WebHoverOverlay`. It
operates directly on the DOM node representing the hovered token, but
that wouldn't work for CodeMirror.

I briefly considered extending the component with an additional prop to
notifier the parent/caller when the component determined that "click to
go to definition was available". But that would only make the component
more "messy".

Instead I noticed that the whole "click to go to definition" logic inside
`WebHoverOverlay` actually operates on data passed into this component and
strictily speaking didn't have to reside inside that component at all.
But since I don't want to refactor the existing blob view either, I moved as
much logic as possible into a reusable function and now use that
function inside the CodeMirror hovercard class.

How does that solve the problem of showing the pointer cursor?
Until now the hovercard class would always register a click event
handler and whether or not this handler had any effect dependent on what
`WebHoverOverlay` did.
But now we can determine whether or not "click to go to ..." is
available or not *before* we register the click handler. So now the mere
presence of the handler indicates that "click to go to ..." is indeed
available and we can use that add an additional CSS class to the token
decoration to show a pointer.