Run the tests for client/common
For this to work, we needed to reference the client/shared/dev stuff
explicitly. You can't do that from the root, because it won't be able
to find those files, so the right way of doing this is to add the
copy_to_bin rule to client/common/BUILD.bazel and then reference that
target from client/shared/BUILD.bazel.