Remove contributions API reliance of new references panel (#42164)
There are currently two parts of code intel that rely on the old
extensions contributions API:

1. The "Mix precise and search-based results" button in the references
   panel.
2. The buttons shown on hover tooltips

This PR removes the reliance on the extensions API for 1. by
implementing the button as a React component.

Since we want to still support the possibility to enable legacy
extensions, the `TabbedPanelContent` still queries the contributions API
if it is present. If the same action is found, it will not be rendered
twice though. When this refactoring is complete, contributions won't be
loaded anymore as `extensionsController` will become `null`.