Fix handling of exact matches in the fuzzy finder
Previously, copy-pasting the exact file path into the fuzzy finder
returned matches for all files in the repository. This commit fixes the
behavior so that only the exact match is highlighted.

For some reason, I was not able to reproduce the browser behavior in our
tests. We even have a test case for exact matches and it passes with
correct behavior before this PR. I had implemented a fix for the same
bug in the past, and added a test case to catch regressions, but
something changed recently where the tests still pass but the behavior
in the browser regressed.  Either way, the behavior in the browser works
as expected now when you copy-paste