search: compute zoekt SearchOptions in one call (#42425)
At every callsite we computed an opaque k value and then called another
function to generate the search options. We can simplify this into a
single API.

This change was motivated since I wanted to adjust this logic and
potentially add a new parameter. So took the opportunity to simplify the
API.

Test Plan: go test