graphqlbackend: use StableSort for InvitableCollaborators (#42557)
Our tests do not specify a date, which means for the initial sort all
elements are regarded as equal. Go changed the implementation of sort in
go1.19, which means a test for this function started to fail even though
the sort was still correct. Switching to StableSort not only makes the
algorithm deterministic across sorting algorithms, but it also seems
like a better tie breaking condition (respect initial order)

Test Plan: go test