search: tune ResultCountFactor
This disentangles SearchOpts and ResultCountFactor and updates the
limits slightly.

- ResultCountFactor is now independent of the requested match count
and only depends on the number of repos to be searched.
- SearchOpts is now responsible to scale `k` and `MaxWallTime` depending
on the requested match count.

We also fix a small bug where we scaled MaxWallTime even for stream
searches with default count.

Note:

For streaming search we ignore TotalMaxMatchCount, MaxDocDisplayCount in
Zoekt and rely on Sourcegraph cancelling the context once the limit is
reached.

TotalMaxImportantMatch has no effect in any case.