gitserver: correctly join path to CloneFromShard (#42564)
filepath is the incorrect package to use for joining a path onto a
URL (use path package). Additionally, path.Join is not correct to use on
a URL since it is supposed to operate on a path, not a URL. In go1.19
path join additionally does cleaning. This means things like "http://"
will change into "http:/", leading to the feature breaking.

This commit introduces a vendored version of go1.19s url.JoinPath in our
url wrapper package and uses that in gitserver.

Test Plan: go test works on go1.19 and go1.18