[POC] Codemirror blob: Add toggle to control overriding CMD/CTRL-f
This commit adds a checkbox for controlling whether or not we override
Cmd/Ctrl-f on the file page.

By default, Cmd/Ctrl-f will search inside the file. Disabling that
feature will remove the override and always show the file search input
because otherwise there woulnd't be a way to show the file search.

I'm using local storage for now because temporary storage adds the
additional complexity of dealing with the loading state.