Move auth check helpers out of cmd/frontend
Those can be useful generally, so moving them out of the frontend command. Also, they're not specific to it. Looking at the updated imports, it has already been used outside, but created a bad dependency on frontend code.

Next step would be to take a generic DB type thus getting rid of the dependency on internal/database, so we can also use that package in there.