frontend: remove health-check POC
Work on this has been deprioritized. Let's remove this code. We can add
it back in the future in case we want to pick it up again.

To future travellers:
The missing piece to a fully functional health checker is something like
a service registry that provides network addresses of all
services (including indexserver) to the health checker.