Fix legacy-extensions by passing settings instead of observable (#42701)
Turns out that all of the `getSetting` calls returned `undefined`
because `context.settings()` always returned `undefined`, because
`newSettingsGetter` didn't receive the `SettingsCascade<Settings>` it
expected but the `platformContext.settings`, which is an Observable.

Why that didn't produce a type error? No idea

How to verify that this is correct? No idea

But I did test this locally and can confirm that now settings are
propagated from my global settings. I added print statements to
`getSetting` in `api.ts` that print the values.