sg config: use two gitserver instances by default
We talked about this in repo-management & IAM and we think we should
increase our dogfood consumption.

This is one step in that direction: most of our customers run mulitple
gitserver instances. This helps us make sure we dogfood that.

ATTENTION: this will lead to some repositories being recloned/reindexed
locally.