Convert executor VM definitions to HCL (#42697)
This finally converts us to hcl, which should make it easier to implement some form of logic (json not so smart :clueless:)

With this PR I introduced a new way to configure the aws regions, and we will publish now to all regions on tagged releases, but only our testenvs for insiders builds. Also made the list exhaustive. World wide executors :)