batches: fix webhook log page responsiveness (#42731)
When the content of the request/response body of a webhook is larger than the width of the MessagePanel it causes the panel to extend out of its container. This makes the page look really bad and out of shape when a node is expanded.

There was also an extra column in the grid-template-column for the WebhookLogPageHeader component, which caused the component's width to expand out of its parent container when the number of errors is >1000. This PR ensures the grid for the header defines the correct number of columns.