frontend: revert part of #42736 to fix asset loading (#42855)
The seemingly unused `init` function removed in #42736 was load bearing,
but not in an obvious way: the global variable it set was used in a
different part of the package when generating asset URLs. Without that
variable, asset URLs are missing their `.assets` section, and are
therefore resulting in requests that Sourcegraph is routing to the repo
handler.

This PR reinstates the `init`, but closer to the actual code that uses
it.

Fixes #42854.