Adds clarity regarding different upgrade paths (#42884)
* Adds clarity regarding different upgrade paths

It is not immediately clear on first visiting this page that there are two upgrade paths. A customer may pick the wrong path, not realizing that it will lead to a complete failure of their instance, requiring reversion and repair. This update aims to clarify that choosing the right upgrade path is vital to a successful upgrade.

* removes superfluous section titles 

also adds emphasis to critical distinguishing markers for each upgrade.