Remove Index:bool from RepoListOptions since it's useless
This was added in anticipation of a feature that didn't arrive: ability
to only index a subset of repositories (by adding a `index boolean`
column to `repo`).

The change here removes the option but preserves the original behaviour:
on dotcom the option was and is ignored, on non-dotcom instances we
return an empty list of repositories if indexing has been disabled (but
this time without hitting the databse).