ranking: stream out as we receive an edge
A small change I noticed when reviewing the PR which added
buildFileReferenceGraph. We don't have to collect all paths before
streaming, instead we can check if we have seen a path before and stream
it out if we haven't.

I did this commit hastily, not sure if I am a fan of the name "gOut" and
leaving the other group called "g".

Test Plan: CI