ranking: serve rank instead of 1-rank (#42922)
In the code we use both rank and 1-rank, which isn't necessary and
causes confusion (at least for me).

Zoekt originaly used 1-rank in its sorting logic for documents. In Zoekt
ranking happened within sorting, so the logic was very confined.

When we copied the ranking logic from Zoekt to frontend we copied the
approach.  However, now that ranking  happens in frontend and sorting in
Zoekt, we should stop using 1-rank and just serve rank which is much
more intuitive.