packages: stop fetching/gitPushDependencyTag if package not found
On S2 and dotcom we have a lot of packages that fail to clone but the
error message that show up like this:

    failed to clone python/asvdb: get clone command: failed to fetch repo
    for file://python/asvdb: command [git branch --force latest v0.4.2]
    failed with output fatal: Not a valid object name: 'v0.4.2'.: exit
    status 128

It's very cryptic. The reason for that is that we've previously
*ignored* when a package wasn't found and then continued to execute the
`Fetch()` method that would then try to `git branch --force latest
<version>` to set the latest tag, etc.

This change here logs the not-found error and also returns it. All the
callers that _should_ ignore an error like this (`fetchRevSpec`) already
ignore errors from `gitPushDependencyTag`.