executor: Use git proxy service for cloning (#42927)
* executor: Use git proxy service for cloning

This is so that we never need to give git any credentials. It will also prevent the credential from ever landing in logs, so even if redaction fails this shall not cause any problem. It also makes sure that no further git remote operations are ever possible after the cloneRepo function returned.

* Fixup test