ranking: send repo rank to zoekt (#42985)
DO NOT ENABLE THIS ON .COM, otherwise this will negatively impact the ranking of freshly indexed repos.

This is behind the same ENV variable (ENABLE_EXPERIMENTAL_RANKING) that the ranking service uses.

If enabled, frontend-internal returns the rank instead of the star-count in the RepoIndexOptions.

Alternatives considered:
We could send the ranking vector to Zoekt and persist it in the shard, maybe as part of rawConfig. However, this would require a slightly bigger change with pretty much the same effect, at least for the demo.