executors: Undo service-ification for now
The service pattern is a great long term goal, but it currently makes the assumption that we can freely use services with each other, and that is not true yet, sadly. For example, it is impossible to import from graphqlbackend because it caused a cyclic dependency. That is required though to extend the API.
To unblock, this will revert that new pattern for now and we can look at it more wholistically later.