executors: Fix env vars with spaces
When we introduced quoting for env vars, we went a bit too far and actually broke env vars with spaces for docker commands that don't run in firecracker.
When we quote the env vars in the docker layer, they will be literally quoted in the value that docker receives, because we use exec.Command there and it handles the arg splitting already.
For firecracker, we actually make the command a shell script that we run so there we have to encode it right.