executors: Make src upload step in codeintel a docker step
We want to get rid of special casing src-cli. This is the first step towards that. src-cli will still be required for batch changes for the time being but I am working on resolving that as well, but it'll be more involved.
To make sure that customers using custom registries can still use this just fine, I added 2 new site settings that make this image configurable. This will also allow to ship a src hot fix to a specific customer, if src is version compatible with that backend.