minio: Untag latest before retagging it
Docker is currently running an experiment where containerd is used to pull images, which is an experimental flag in Docker for Mac for now but might become the standard. It doesn't seem to handle tagging images 100% the same way as docker does though, as it doesn't allow using an already used tag.
This fixes it by untagging the image first before tagging it again.
With this fix, sg start can start up just fine.

If we don't want to fix this upstream that's ok, just figured it might be good future proofing as I just played around with that new flag. Let me know what you think :)