CodeMirror blob: Only select lines on main mouse button click; don't show hovercards when selecting text (#43117)
The current implementation selects lines when any mouse button is clicked which interferes with people's use of forward and back buttons.

And while I was looking at mouse events I also updated the hovercard code to not show/hide hovercards when the user moves the mouse and presses a mouse button (e.g. when selecting text).