search input: Reduce "flicker" of query suggestions
This commit makes use of CodeMirror's `validFor` option to reduce the
"flickering" of search suggestions. If set, then CodeMirror doesn't
query the source anymore as long as the `{from, to}` range matches the
provided regular expression.

I've only enabled this for filter names and filers with static values
because it's not clear how filters with both static and dynamic values
should work in this regard.