Fuzzy finder: refactor and implement Figma design
This is the first step towards implementing the design from
https://www.figma.com/file/E0vD9UduSp2liVyDb7dplw/Fuzzy-finder-UI-improvements?node-id=1714%3A3559

While working on the code, I was increasingly unhappy with the
convoluted state management inside `useFuzzyState`, which has caused app
freezes (infinite loops between cyclic hooks).

The new state management inside `useFuzzyState` is significantly
simpler to reason about and it should be much easier to add new
functionality to this new codestructure.