gitserver: add command LFSSmudge
This adds a new command which will run the contents of files through
git-lfs-smudge. Smudge is the terminology used by git filters for
converting a file from what is stored in git into what you want in your
working copy. So what this command will do is fetch the file from LFS.

If the file is not an LFS pointer, git-lfs-smudge ignores it and passes
it on. We use the same behaviour in our implementation, but we bypass
needing to pass the large file as stdin to gitserver.

We needed to fix LocalGitCommand.StdoutReader since previously it was
returning combined output (instead of stdout). We also add a utility
which ensures we run against bare clones since this is important to test
for git-lfs.

Note: this is still missing other pieces before it can be used properly.
In particular we need infrastructure to monitor and prune the LFS cache.
Additionally I need to test more how authorization works in filling the
LFS cache, which means we may need to change this implementation up a
bit.

Test Plan: added a test