executors: Implement stores for secrets
This PR is the first of a series of PRs to implement secrets in executors. It introduces 2 new stores that together form one unit:
- ExecutorSecrets: A store that holds all defined executor secrets and access to those, plus encryption
- ExecutorSecretAccessLogs: A store that holds all access log events, these are created automatically when the secret value of the secret is ever read

The secrets store uses both namespace checking in-app for nice error messages and additionally does authz on the SQL layer, so that never a secret is returned to a user that they're not supposed to see.

In the next PR, these stores will be used to implement a resolver layer to expose basic CRUD functionality for these.