executors: Explicitly use internal actor for API (#43326)
From an API perspective, we treat executors as internal and the executor secret is an internal actor access token. This is no change to the before state, just makes sure that we don't forget to set the HTTP header (which actually was forgotten on some client calls).
Eventually, this should be removed when we have more restrictive executor tokens.