executors: Support bash syntax when available (#43046)
Executors for now ran commands as scripts in sh only, this adds support for bash as well. This is something that we supported in batch changes forever due to customer feedback.
In batch changes, we used an approach where we would run the container with `/bin/bash mktemp` and then `/bin/sh mktemp` to determine which shell is available. I think this approach is actually nicer, because it
- respects the PATH setting in the container for bash
- doesn't require up to 2 additional container boots before the actual command can be run