all: include git-lfs in gitserver docker images
We add the optional dependency git-lfs to any image gitserver may run
on. This is still an experimental feature, but is a requirement when
enabled.

Additionally we include it in the nix devenv since its cheap to add.

Note: git-lfs is a relatively small go binary and does not require
specific versions of git. Additionally git will not automatically use
it, so its presence will not change behaviour.

Test Plan: CI docker image dry run