Parallelize inference jobs (#43424)
Currently, we are only inferring configs for 360k repos every 12h on dotcom. To raise this number, we want to try if it gets better if these are not fully sequentially inserted, starting with a default concurrency of 16 but added a variable to tweak it more before we settle on a final recommended value for customers.