search: improve match grouping behavior when `by-zoekt-ranking` is enabled (#43378)
* fix logic in ZoektRanking.ts which decides whether to merge a MatchItem into a match group
After updating ZoektRanking.ts to work with multiline matches instead of single-line matches, the behavior of isMatchWithinGroup() changed slightly which resulted in matches with overlapping or adjacent context not being merged into a single match group. This produced unintuitive search results when "searchClientResultRanking" was set to "by-zoekt-ranking" in site config. This commit changes isMatchWithinGroup() to return true if a MatchItem has adjacent or overlapping context with a match group.

* add tests and test data for multiline match results

* wip

* wip