accesslog: RFC using conf.Cached for IsEnabled
This is a simplification of the current code I read. However, it does
not have the same behaviour around logging (it only logs on startup).
This commit exists as a discussion point, and can be completely
abandoned. If it is liked, it needs more work since the tests now fail
since they assert on logging.

conf.Cached is a generic function which takes in a function and then
returns a function which wraps it. Its behavior is:
- on first call it just calls the passed in function and returns.
- then it starts to watch config, each time the config is changed the
function is called again and the new return value is returned.

This is quite a neat pattern, but likely needs some changes to suit this
use case. In particular behaviour around the value changing may be a
requirement to be logged.