WIP accesslog: use log.Field instead of map for meta
This has better memory usage, natively supports slices, does less work
if disabled and is the final target for our logs anyways. As log becomes
more deeply embedded in our codebase, I imagine this will be the more
convenient field to use in logs, traces and accesslog.

I am unsure if we should proceed with this, so this is just a mini PR
implementing the idea, but not tested at all. IE give feedback :)