Add `zoekt_repos` table, populate it in worker, add endpoint for Zoekt to update (#43289)
This adds the `zoekt_repos` database table that mirrors what the
`gitserver_repos` table does:

A row is added to `zoekt_repos` when a new row is `INSERT`ed into
`repo`. `index_status` is `"unindexed"`.

A worker then talks to Zoekt every 30min to update this table and the
index status of each repository.

Since that's too slow to show it to the user ("wait, my repo still
hasn't been indexed?!"), there's a new endpoint for Zoekt to talk to. In
another PR I have not yet opened, I added a method to
`zoekt-indexserver` that causes it to send a request to that endpoint
and update the index status as soon as it's done indexing.

All of that will allow us to efficiently query which repositories are
indexed and what the overall progress of indexing on the instance is.

What will come later:

* using this table to display counts in site-admin area (and filtering of repositories)
* a new status message that shows indexing progress
* sourcegraph.com support. On dotcom we don't index every repository
  (which is what we do on a customer index, if indexing isn't turned
  off), so the querying of the table is a bit more complicated. Since
  there's a lot of noise on dotcom, I defered on that.