remove extension preloading
Extensions were deprecated in Sourcegraph 4.0. Preloading extensions is a performance optimization that should only be invoked when the deprecated extensions feature is enabled, but it was (incorrectly, but harmlessly) being invoked always. This is harmless, but it makes it confusing to debug problems and results in needless HTTP requests and script executions. This fix removes preloading entirely, which slightly degrades performance for instances using the deprecated extensions feature, but since this entire feature is going away, I think it makes more sense to just remove preloading entirely rather than implement a more nuanced fix.