docker-images: update libxml2 for CVE-2022-40303 CVE-2022-40304
This only updates the base image, we still will need to follow up and
update alpine references once landed.

We also fix the upgrade calls in postgres since they didn't quote the
">" char which means shell redirection. This is probably why it was
written as muliple commands instead of one since it would of been a
shell error.

Test Plan: docker image builds on CI