search: collect and flush ranking (#43604)
This is our new experimental ranking path which overcomes the limits of
bucketing by stars. It is feature flagged behind "search-ranking", so
can be turned on per user. It relies on the recently shipped
FlushWallTime feature in zoekt.

When turned on we will effectively wait 500ms to collect candidate
results. We will then rank and truncate those results based on combining
Zoekts query ranking and our new #ranking work.

Note: This is all experimental #ranking work so right now we hardcode a
bunch of assumptions into the code. As we gain emperical understanding
we will introduce ways to tune / more sophisticated mechanisms to pick
values like FlushWallTime. We have some prometheus metrics to understand
if FlushWallTime needs tuning.

Test Plan: Minimal functional changes unless feature flag is turned on.
So intend to rely on unit tests to ensure we didn't break existing code
paths. Otherwise for now will ne relying on manual validation (which
will be comparing searches).