move legacy codeintel to enterprise (#43578)
* fix error in OSS build when hovering tokens

Fixes:

```
TypeError: Cannot read properties of undefined (reading 'fields')
    at API.hasStencils (api.ts:579:84)
```

Related to https://github.com/sourcegraph/sourcegraph/pull/43458, but that did not fully fix the issue (seemingly because the GitBlobLSIFData type is not even introspectible on OSS because LSIF is not available in OSS).

* remove/unexport local defs

* move useCodeIntel hook to /enterprise/ tree

* move legacy codeintel to enterprise

The search-based code intelligence is an enterprise-only feature. This change clarifies it by showing a nicer message instead of an error if it's attempted to be used in the OSS build. It also cleans up how the enabled-or-not flag is determined in code.