Fuzzy finder: with filename first
Previously, the full file path was listed making it sometimes difficult
to find the filename since could disappear behind the horizontal scroll
for long filenames (esp. common when searching globally).

Now, the filename is displayed first and the directory is displayed
after. This was somewhat tricky to implement because we want to allow
the user to query the full path (directory+filename).