internal/database: Order by ID only if no order by args are present
Previously, we made a change to _always_ order by ID in this PR:
https://github.com/sourcegraph/sourcegraph/pull/27026

The reasoning back then was to speed up the query which would take >30
seconds on dotcom. It turns out we are back at square one after about
a year. On further investigation, the change back then made sense
because we used to have a clause to filter against
external_service_repos.org_id in the "queryConds". However, we have
since stopped the usage of this column for filtering repos.

As a result, now that the set of repos is not filtered down by one
specific org_id, the resulting set is huge on dotcom and the sorting
across two dimensions (name, id) ends up taking ~7s on its
own. Additionally it no longer uses the indexes on name as a result.

But sorting by just name takes <150ms with the same query as it is now
able to use the repo_name_unique index instead of a sequential scan.

The performance gain is huge (~280ms vs 30s) for the repositories
GraphQL query on dotcom. Example:

```
query {
  repositories(first: 1) {
    nodes {
      name
    }
  }
}
```