ranking: introduce DocumentRankVersion
This field allows us to indicate when the ranks have changed, which
forces us to re-index. This field is currently the updated at timestamp
which will change as we update ranks for a repository.

This commit is paired with a change in Zoekt that reads this field when
set. Now when ranking is enabled in frontend it will be enabled in
zoekt. So when this PR lands (with the zoekt one) we will start to store
the page ranks in zoekt for use at query time. However, using those
ranks fully is still behind a feature flag for query time.

Test Plan: Mostly relied on existing unit tests and unit/manual testing
within the Zoekt repo. Otherwise this is behind an envvar feature flag
for sourcegraph.com