src-expose feature: ignore files in source directory
By default, `src-expose` will sync and serve all content
in a source directory.
When using `src-expose` on local repositories from other SCM tools,
like Subversion, the SCM local database and control files should
be ignored to avoid Sourcegraph indexing those files.
For most setups, build and dependency directories should be ignored also.

Add to the config definition an `ignore` property on `dir` that is a
list of `gitignore` patterns identifying items in the source directory
to exclude from the sync and serve.

When using the `ignore` list, there is no need to create a `.gitignore`
file in the source directory, as shown in the `example.yaml`
configuration file.

The `ignore` list is written to the file `$GIT_DIR/info/exclude`,
where `$GIT_DIR` is the `.git` dir in the destination directory,
so the source directory is not touched.

See the `git` documentation about the `$GIT_DIR/info/exclude` file
and `gitignore` patterns: https://git-scm.com/docs/gitignore