graphqlbackend: use injected gitserver.Client (#43791)
We had a few call sites I found in graphqlbackend which still called
NewClient when the struct itself already had a client in it. I checked
the construction to ensure we always passed in, but in theory there is a
risk of a panic in prod here. Low likelyhood though.

Test Plan: go test