authz: Update handling of directories in sub-repo filtering
We want to include all parent directories of an include match. We used
to store this in a separate structure but this would not take into
account rule order. Instead, we now insert an include rule for each
parent directory along with the rule itself into our normal rule list.

In the case of rules that begin with wildcards we can't generate a
finite list of parent directories so instead we insert a rule that
matches any directory, effictiley any path ending in '/'.

This change also works around an edge case in the glob package we use
where a rule like `/**/*.java` would not match java files in the root of
the tree.